+++ src/Jellyfin.LiveTv/Timers/TimerManager.cs
@@ -7,7 +7,7 @@ using System.IO;
 using System.Linq;
 using System.Threading;
 using Jellyfin.Data.Events;
-using Jellyfin.LiveTv.EmbyTV;
+using Jellyfin.LiveTv.Recordings;
 using MediaBrowser.Common.Configuration;
 using MediaBrowser.Controller.LiveTv;
 using MediaBrowser.Model.LiveTv;
